# 🍪 Implement YouTube Cookie Authentication