… COOKIE-AUTHENTICATION-IMPLEMENTATION-REPORT.md
… YOUTUBE-COOKIE-AUTHENTICATION-GUIDE.md
… backend/build/services/streamingService.js
… test-youtube-cookie-auth.js